Coerce event_id to string
The event_id field will eventually be mapped as a keyword type in
Elasticsearch, so casting as a string makes sense.

Additionally,
Logstash initially casts the field as a number, so conditionals
comparing values in quotes fail without the convert being applied.

In example:

\# Conditional evaluates to false
if [event_id] == "20250" {
    ...
}

\# Conditional evaluates to true
mutate {
    convert => { "event_id" => "string" }
}

if [event_id] == "20250" {
    ...
}